backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
torch==2.7.0
torchvision==0.22.0
pillow==11.2.1
langgraph==0.4.7
langchain-google-genai==2.1.4
langchain-community==0.4.1
langchain-core==0.3.51
tavily-python==0.7.3
requests==2.32.3
pydantic==2.11.3
